Main: move MinGW support to D3D9 & D3D11 RenderSystems
they were the only consumers. Slightly duplicates code. Slightly
simplifies build.
