+++ RenderSystems/Direct3D11/include/OgreD3D11Prerequisites.h
@@ -31,7 +31,7 @@ THE SOFTWARE.
 
 
 #include "OgrePrerequisites.h"
-#include "WIN32/OgreMinGWSupport.h" // extra defines for MinGW to deal with DX SDK
+#include "OgreMinGWSupport.h" // extra defines for MinGW to deal with DX SDK
 #include "OgreComPtr.h"       // too much resource leaks were caused without it by throwing constructors
 
 #include "OgreException.h"
