+++ RenderSystems/Direct3D11/include/OgreMinGWSupport.h
