+++ RenderSystems/Direct3D9/include/OgreMinGWSupport.h
@@ -1,0 +1,54 @@
+/*
+-----------------------------------------------------------------------------
+This source file is part of OGRE
+    (Object-oriented Graphics Rendering Engine)
+For the latest info, see http://www.ogre3d.org/
+
+Copyright (c) 2000-2014 Torus Knot Software Ltd
+
+Permission is hereby granted, free of charge, to any person obtaining a copy
+of this software and associated documentation files (the "Software"), to deal
+in the Software without restriction, including without limitation the rights
+to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+copies of the Software, and to permit persons to whom the Software is
+furnished to do so, subject to the following conditions:
+
+The above copyright notice and this permission notice shall be included in
+all copies or substantial portions of the Software.
+
+THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+THE SOFTWARE.
+-----------------------------------------------------------------------------
+*/
+#ifndef __Win32MinGWSupport_H__
+#define __Win32MinGWSupport_H__
+
+// This header provides a compatibility layer for some MSVC symbols
+// for MinGW. They may be needed when using MSVC libraries, most
+// notably the DirectX SDK.
+
+#ifdef __MINGW32__
+#include <stdint.h>
+
+// define a number of symbols MSVC uses for annotation.
+#include <specstrings.h>
+#define __in_z
+#define __in_z_opt
+#ifdef __MINGW64_VERSION_MAJOR
+#   define __in
+#endif
+#define UINT8 uint8_t
+#define WINAPI_INLINE inline
+#ifndef __uuidof(Object)
+#   define __uuidof(Object) IID_##Object
+#endif
+
+#endif
+
+
+#endif
