+++ RenderSystems/Direct3D9/src/OgreD3D9Plugin.cpp
@@ -31,7 +31,7 @@ THE SOFTWARE.
 
 #ifdef __MINGW32__
 extern "C" {   
-#include "WIN32/OgreMinGWSupport.h"
+#include "OgreMinGWSupport.h"
 void _chkstk();
 void _fastcall __security_check_cookie(intptr_t i);
 }
